(CDAP-5684) Don’t stop Worker in the test.
Also speed up client tests by
making most client tests not to have explore enabled.
